Add Server Side event handling
This commit adds a server-side event to the API[1] via the new
metadataHandlers.ts file. A server-side event is a unidirectional stream
that can send data to a client based on a server-side trigger, in this
case a post request to the metadata endpoint. This is done to take
advantage of spinitron's metadata push interface, which sends a post
request containing a dearth of useful information about the latest
logged spin every time a new spin is logged. In doing this, we are able
to receive up-to-date information on the current stream without having
to worry about repeatedly querying the spinitron API.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/Using_server-sent_events