Remove SSE from metadata routes
This commit dumps the sse approach from previous commits
in the metadataHandlers in favor of a simpler caching approach.
Now, the getMetadata route will just return the most recently pushed
song, which is set by spinitron calls. To update, we will just
be polling repeatedly on the front-end every 5 seconds or so.