Add Spins Caching
This commit adds another cache to the API to store current spins.
Notably, this cache utilizes another cache, the metadata cache,
to read the current playlist id. This is the first time cache variables
are being passed around, but future commits should take advantage of this
functionality. Namely, writing in handlers that return custom combinations
of the various caches to minimize the number of requests to our proxy.