Refactor Caches into Stores
This commit refactors the existing caches into the new 'stores'
directory, and renames them accordingly. The idea here is to separate
the caches from the handlers, as they function independent of the api
request handling. This improves readability for the handler files, and
adopts a pattern that should be helpful when defining stores moving
forward. Furthermore, it supports infrastructure that would require
multiple stores (such as the current Spins store, which requires the
context of the current playlist id, which it can retrieve from the
currentPlaylistStore).