Refactor Current Route
This commit gets rid of the separate getCurrentSpins and
getCurrentPlaylist routes and combines them into one route, '/current',
which returns the stored information for both fields. The separate
'spins/current/spins' and '/playlists/current/playlist' routes were both
ugly to look at and meant that we would need to poll our api twice to
get the information, when we are perfectly able to provide it in one
call.