Add Metadata to Current
This one-line commit adds the metadata (as recieved from spinitron's metadata push api) to the current GET endpoint, which will give us access to the metadata data on request. This was done specifically to ensure that we have access to the current dj name when we poll for current data, as that would otherwise be another request.