Add Personas Store and Error Handling
This commit adds a personas store which is used to cache persona data at
a regular interval, every hour. This is slightly more involved than the
other stores as it needs to query multiple pages of data (or at least
potentially multiple) as the number of personas we have can
grow. However, since that value should never be more than a thousand (at
VERY most) it seems sustainable to just query every page once every hour
or so. Interval length can be increased if needed. It also fleshes out
the error handling in the repository to actually include useful error
messages, instead of just generic errors. AND, now the error handling
middleware is actually functioning, which I had not realized it was not
due to the middleware import order.