Force Persona Expansion in Shows routes
This commit adds a 'expand=personas' parameter to all shows routes,
which makes them include the personas as an array in their response. It
also drops the url parameters pattern that allows for endpoints to
receive a variable number of search queries (it drops this pattern in
the shows handlers). These are being dropped here principally because
they didn't seem to work, but also because ultimately we should handle
all parameters to spinitron in the API, not on the frontend. Some future
commit should remove all those url search parameter setups in the other
handlers.