Add configurable show limit for show fetch
This commit adds a configurable variable for our api to dictate the
number of shows returned in a single show call. This was originally
handled by the front-end when it was proxying query parameters to
spinitron, but since we have moved away from that, we want to now handle
retrieving all shows in our database on our api end.